+++ src/main/java/com/manpower/backendProject/repositories/TokenRepository.java
@@ -9,14 +9,14 @@ import java.util.List;
 import java.util.Optional;
 @Repository
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     @Query(value = """
       select t from Token t inner join User u 
       on t.tokensUser.id = u.id 
       where u.id = :id and t.revoked = false
       """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(long id);
 
     Optional<Token> findByToken(String token);
 }
